+ Pages, 'App' cleanup
Cleaned 'App' component.
Moved all the references to 'References' component.
Added dummy 'Search' and 'Study' pages.
