+++ src/pages/Search.js
@@ -1,5 +1,12 @@
+import { SearchField } from 'features';
+
 const Search = () => (
-  <p>This is a search page</p>
+  <>
+    <p>This is a search page</p>
+    <SearchField />
+  </>
 );
+
+Search.displayName = 'Search Page';
 
 export default Search;
