+++ src/pages/Study.js
@@ -2,4 +2,6 @@ const Study = () => (
   <p>This is a study page</p>
 );
 
+Study.displayName = 'Study Page';
+
 export default Study;
